+++ Sources/BraintreePayPal/BTPayPalVaultRequest.swift
@@ -12,17 +12,13 @@ import BraintreeCore
     /// Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`.
     public var offerCredit: Bool
 
-    /// Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
-    public var userAuthenticationEmail: String?
-
     // MARK: - Initializer
 
     /// Initializes a PayPal Native Vault request
     /// - Parameter offerCredit: Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`.
     /// - Parameter userAuthenticationEmail: Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
-    public init(offerCredit: Bool = false, userAuthenticationEmail: String? = nil) {
+    public init(offerCredit: Bool = false) {
         self.offerCredit = offerCredit
-        self.userAuthenticationEmail = userAuthenticationEmail
 
         super.init(hermesPath: "v1/paypal_hermes/setup_billing_agreement", paymentType: .vault)
     }